requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
Pillow==9.5.0
python-multipart==0.0.20
torch==2.2.2
torchvision==0.17.2
numpy==1.24.4